pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cutsimulator"
version = "0.1.0"
description = "Kubernetes workload simulator"
readme = "README.md"
requires-python = ">=3.8"
authors = [{name = "CUT"}]
license = {text = "MIT"}
dependencies = [
    "kubernetes>=25.3.0",
    "PyYAML>=5.4",
    "numpy>=1.21.0",
    "torch>=1.13.0",
    "pandas>=1.3.0",
    "pettingzoo>=1.24.0",
    "gymnasium>=0.26.0",
    "typing_extensions>=4.0.0"
]

[project.scripts]
cluster-controller = "scripts.cluster_controller:main"
simulation-controller = "scripts.simulation_controller:main"
training-controller = "scripts.training_controller:main"
